Implemented on demand feature.
Go get the sources (in relay mode) at startup then disconnect.
Reconnect when clients connect and ask for the source.
Disconnect when there is no more listeners after a timeout.
Setting can be done globally or by mount point (or both).
On demand is disabled by default if not present in configuration.